+++ src/main/java/devs10x/mtg/devs10x_mtg_deckbuilder/entity/Card.java
@@ -18,5 +18,5 @@ public class Card {
     private String apiId;
     private String name;
     private String manaCost;
-    private Double cmc;
+    private Integer cmc;
 }
